requirements.txt
kafka-python-ng
pyspark
vaderSentiment
pymongo
streamlit
pandas